+++ lib/glimmer/libui/control_proxy/area_proxy.rb
@@ -37,6 +37,7 @@ module Glimmer
         end
         
         LISTENERS = ['on_draw', 'on_mouse_event', 'on_mouse_move', 'on_mouse_down', 'on_mouse_up', 'on_mouse_drag_start', 'on_mouse_drag', 'on_mouse_drop', 'on_mouse_crossed', 'on_mouse_enter', 'on_mouse_exit', 'on_drag_broken', 'on_key_event', 'on_key_down', 'on_key_up']
+        
         LISTENER_ALIASES = {
           on_drawn: 'on_draw',
           on_mouse_moved: 'on_mouse_move',
@@ -48,6 +49,7 @@ module Glimmer
           on_mouse_exited: 'on_mouse_exit',
           on_drag_break: 'on_drag_broken',
         }
+        
         SHIFTED_KEY_CODE_CHARS = {
           '`' => '~',
           '1' => '!',
@@ -72,7 +74,6 @@ module Glimmer
           "\\" => '|',
         }
 
-        
         include Glimmer::FiddleConsumer
         include Parent
         prepend Transformable